Moved tests in separate file
